requirements.txt
opencv-contrib-python
tensorflow
tensorflow-addons
torch
torchvision
pandas
seaborn
Pillow
numpy
matplotlib
PyYAML
tqdm
